+++ CHANGELOG.md
@@ -1,0 +1,24 @@
+### Changelog
+
+All notable changes to this project will be documented in this file. Dates are displayed in UTC.
+
+Generated by [`auto-changelog`](https://github.com/CookPete/auto-changelog).
+
+#### [v0.2.1](https://github.com/sambacha/mempool-shark/compare/v0.1.1...v0.2.1)
+
+> 7 September 2021
+
+- refactor(service): upgrade deps, cleanup schemas and rework build scripts [`d3d7f41`](https://github.com/sambacha/mempool-shark/commit/d3d7f41068a2026715ecb871b8a80fa4ae2b5d99)
+- lint(format): apply [`6d06ab3`](https://github.com/sambacha/mempool-shark/commit/6d06ab3c88ea03b0bb84b8eb508169e224098c80)
+- ci(actions): nodejs [`35c340a`](https://github.com/sambacha/mempool-shark/commit/35c340a0ec3a34ef74dcb057a9fba636cf76aa44)
+
+#### [v0.1.1](https://github.com/sambacha/mempool-shark/compare/v0.1.0...v0.1.1)
+
+> 17 August 2021
+
+- feat(docs): init [`3652544`](https://github.com/sambacha/mempool-shark/commit/36525443fa896045c8b8a52ea634870484591bb1)
+- Release 0.1.1 [`a0b8406`](https://github.com/sambacha/mempool-shark/commit/a0b84060254e5b6331e93e68d4bb26b1a71bb118)
+
+#### v0.1.0
+
+> 17 August 2021
