+++ src/listener-confirmation.ts
@@ -2,9 +2,9 @@ import {startMongo, models} from './utils/mongo/config';
 import {escan2, MAIN_WS_PROVIDER, providersForLC} from './utils/web3/providers';
 import {_log, timeout} from './utils/configs/utils';
 import {getPendingTxResponse} from './utils/web3/getTransactions';
-import {proccessPending as pendingTx_uni_sushi} from './pipeline/venues/pending';
+import {proccessPending as pendingTx_uni_sushi} from './pipeline/protocols/pending';
 import {pendingToConfirm, trashToconfirm} from './utils/mongo/saveConfirmed';
-import {Document as _doc} from 'mongoose';
+import { Document as _doc } from 'mongoose';
 
 const {txM, g} = models;
 const {whales} = g;
